requirements.txt



alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.11.1
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
coverage==7.13.5
cryptography==46.0.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.104.1
googleapis-common-protos==1.73.1
greenlet==3.3.2
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
opentelemetry-api==1.40.0
opentelemetry-distro==0.61b0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-asyncio==0.61b0
opentelemetry-instrumentation-click==0.61b0
opentelemetry-instrumentation-dbapi==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-grpc==0.61b0
opentelemetry-instrumentation-httpx==0.61b0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-psycopg2==0.61b0
opentelemetry-instrumentation-redis==0.61b0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-instrumentation-sqlite3==0.61b0
opentelemetry-instrumentation-starlette==0.61b0
opentelemetry-instrumentation-threading==0.61b0
opentelemetry-instrumentation-tortoiseorm==0.61b0
opentelemetry-instrumentation-urllib==0.61b0
opentelemetry-instrumentation-urllib3==0.61b0
opentelemetry-instrumentation-wsgi==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
prometheus-client==0.19.0
protobuf==6.33.6
psycopg2-binary==2.9.9
pyasn1==0.6.3
pycparser==3.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
redis==5.0.1
requests==2.33.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
structlog==23.2.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==1.17.3
zipp==3.23.0
